feat: add javadocs and relocate OntopusRequest to api model package
