+++ core/src/main/java/cz/lukaskabc/ontology/ontopus/core/rest/controller/PublicDcatController.java
@@ -5,6 +5,7 @@ import cz.lukaskabc.ontology.ontopus.core_model.config.OntopusConfig;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
 
+/** Configures {@link ResourceController} with {@link DcatResourceService} to serve internal DCAT model */
 @Controller
 public class PublicDcatController extends ResourceController {
     public PublicDcatController(
